+++ src/main/CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.3] - 2016-01-30
+### Fixed
+- Ender Converter ignoring other Converters on signal validation
+- Ender Converter not notifying other Ender Converters of change
+
 ## [1.0.2] - 2016-01-30
 ### Fixed
 - Converter signal transmission/reception